add threshold for pixel match and also shift back to original structure
