+++ .circleci/config.yml
@@ -305,6 +305,8 @@ jobs:
               export EMSCRIPTEN=$HOME/emsdk/fastcomp/emscripten
               nvm use v11.9.0
               ./build_js.sh
+              # TODO: Later fix this in the cmake build step
+              cp -v build_js/Release/bin/hsim_bindings.* build_js/esp/bindings_js/
       - run:
           name: Run sim benchmark
           command: |
@@ -313,17 +315,8 @@ jobs:
               . activate habitat; cd habitat-sim
               python examples/example.py --scene data/scene_datasets/habitat-test-scenes/van-gogh-room.glb --silent --test_fps_regression $FPS_THRESHOLD
       - run:
-          name: Run JavaScript server
-          background: true
-          command: |
-            cd habitat-sim
-            . ~/.bashrc
-            nvm use v11.9.0
-            npm start
-      - run:
           name: Run JavaScript tests
           command: |
-            sleep 5
             cd habitat-sim
             . ~/.bashrc
             export PATH=$HOME/miniconda/bin:$PATH
