+++ src/esp/bindings_js/tests/test_utils.js
@@ -1,4 +1,36 @@
+import http from "http";
+import finalhandler from "finalhandler";
+import serveStatic from "serve-static";
 import puppeteer from "puppeteer";
+
+export async function createServer() {
+  const serve = serveStatic("./");
+  const server = http.createServer(function (req, res) {
+    var done = finalhandler(req, res);
+    serve(req, res, done);
+  });
+
+  await new Promise((resolve, reject) => {
+    const startServer = () => {
+      server.once("error", e => {
+        if (e.code === "EADDRINUSE") {
+          server.close(startServer);
+        }
+      });
+      server.listen(4004, "localhost", err => {
+        if (err) {
+          reject("Failed to listen on port 4004");
+        } else {
+          const address = server.address();
+          console.log(`Listening on http://${address.address}:${address.port}`);
+          resolve();
+        }
+      });
+    };
+    startServer();
+  });
+  return server;
+}
 
 export async function getBrowserAndPage(url) {
   const browser = await puppeteer.launch({
@@ -13,6 +45,26 @@ export async function getBrowserAndPage(url) {
   return { browser, page };
 }
 
-export function getURL(path) {
-  return `http://localhost:4004/${path}`;
+export async function getServerAndURL(path) {
+  const server = await createServer();
+  const address = server.address();
+  const port = address.port;
+  const url = `http://localhost:${port}/${path}`;
+  return {
+    server,
+    url
+  };
+}
+
+export async function closeBrowserAndServer(browser, server) {
+  browser.close();
+  await new Promise(resolve => server.close(resolve));
+}
+
+export async function generateImage(path) {
+  const { server, url } = await getServerAndURL(path);
+  const { browser, page } = await getBrowserAndPage(url);
+  const screenshot = await page.screenshot();
+  closeBrowserAndServer(browser, server);
+  return screenshot;
 }
