+++ src/esp/bindings_js/tests/viewer_demo.visual.test.js
@@ -1,11 +1,15 @@
-import { getURL, getBrowserAndPage } from "./test_utils.js";
+import {
+  getServerAndURL,
+  getBrowserAndPage,
+  closeBrowserAndServer
+} from "./test_utils.js";
 import { toMatchImageSnapshot } from "jest-image-snapshot";
 
 expect.extend({ toMatchImageSnapshot });
 
 test("viewer rendering should match the snapshot", async () => {
   jest.setTimeout(120000);
-  const url = getURL(
+  const { server, url } = await getServerAndURL(
     "build_js/esp/bindings_js/viewer.html?scene=skokloster-castle.glb&useDefaultEpisode=true"
   );
   const { browser, page } = await getBrowserAndPage(url);
@@ -16,6 +20,10 @@ test("viewer rendering should match the snapshot", async () => {
     'document.querySelector("#status").style.color === "white"'
   );
   const screenshot = await page.screenshot();
-  browser.close();
-  expect(screenshot).toMatchImageSnapshot();
+
+  closeBrowserAndServer(browser, server);
+  expect(screenshot).toMatchImageSnapshot({
+    failureThreshold: 0.1,
+    failureThresholdType: "percent"
+  });
 });
